Move OpacPica-Plugin into separate maven module add target folder to gitignore + remove obsolete Ant-entries
