+++ pom.xml
@@ -63,6 +63,7 @@
     </repositories>
 
     <modules>
+        <module>OpacPica-Plugin</module>
         <module>Kitodo</module>
     </modules>
 </project>
